+++ components/checkbox/test/checkbox.a11y.test.js
@@ -3,9 +3,11 @@ import '@vonage/vwc-formfield';
 import {
 	isolatedElementsCreation,
 	waitNextTask,
-	textToDomToParent,
-	runAxeCore,
+	textToDomToParent
 } from '../../../test/test-helpers.js';
+import { chaiA11yAxe } from 'chai-a11y-axe';
+
+chai.use(chaiA11yAxe);
 
 const COMPONENT_NAME = 'vwc-checkbox';
 
@@ -22,6 +24,6 @@ describe('checkbox a11y', () => {
 		);
 		await waitNextTask();
 
-		await runAxeCore(actualElement);
+		await expect(actualElement).shadowDom.to.be.accessible();
 	});
 });