+++ components/expansion-panel/stories/expansion-panel.stories.js
@@ -5,9 +5,9 @@ import { spread } from '@open-wc/lit-helpers';
 import { argTypes } from './arg-types-expansion-panel.js';
 
 export default {
-  title: 'Components/Atoms/Expansion Panel',
-  component: 'vwc-expansion-panel',
-  argTypes
+	title: 'Alpha/Components/Expansion Panel',
+	component: 'vwc-expansion-panel',
+	argTypes
 };
 
 const Template = args => html`
@@ -19,9 +19,9 @@ const Template = args => html`
   </style>
   <vwc-expansion-panel ...=${spread(args)}>
   <form>
-      A campaign describes a specific use case and details of the messages you will be sending through it. 
+      A campaign describes a specific use case and details of the messages you will be sending through it.
       Such as: sample messages, subscriber opt-in/out, and the associated numbers you will be sending with these messages.
-      
+
       <vwc-textfield label="Selected Brand" value="Vonage Inc." dense readonly></vwc-textfield>
       <vwc-textfield label="Description" dense></vwc-textfield>
     </form>