+++ components/list/src/vwc-list.ts
@@ -1,28 +1,15 @@
 import '@vonage/vvd-core';
+import { debounced } from '@vonage/vvd-foundation/general-utils';
 import { customElement } from 'lit-element';
 import { List as MWCList } from '@material/mwc-list/mwc-list';
 import { style as vwcListStyle } from './vwc-list.css.js';
 import { style as mwcListStyle } from '@material/mwc-list/mwc-list-css.js';
-import { style as styleCoupling } from '@vonage/vvd-style-coupling/vvd-style-coupling.css.js';
+import { style as styleCoupling } from '@vonage/vvd-style-coupling/mdc-vvd-coupling.css';
 
 declare global {
 	interface HTMLElementTagNameMap {
 		'vwc-list': VWCList;
 	}
-}
-
-/* eslint-disable @typescript-eslint/no-explicit-any */
-function debounce(
-	callback: <T>(this: T, ...args: any[]) => void,
-	waitInMS = 50
-) {
-	let timeoutId: number;
-	return function <T> (this: T, ...args: any[]) {
-		clearTimeout(timeoutId);
-		// eslint-disable-next-line @typescript-eslint/no-this-alias
-		const context = this;
-		timeoutId = window.setTimeout(() => callback.apply(context, args), waitInMS);
-	};
 }
 
 /* eslint-disable @typescript-eslint/ban-ts-comment */
@@ -34,8 +21,8 @@ MWCList.styles = [styleCoupling, mwcListStyle, vwcListStyle];
  */
 @customElement('vwc-list')
 export class VWCList extends MWCList {
-	constructor() {
-		super();
-		this.layout = debounce(super.layout);
+	@debounced()
+	layout() {
+		super.layout();
 	}
 }