+++ components/note/test/note.a11y.test.js
@@ -2,9 +2,11 @@ import '../vwc-note.js';
 import {
 	isolatedElementsCreation,
 	waitNextTask,
-	textToDomToParent,
-	runAxeCore,
+	textToDomToParent
 } from '../../../test/test-helpers.js';
+import { chaiA11yAxe } from 'chai-a11y-axe';
+
+chai.use(chaiA11yAxe);
 
 const COMPONENT_NAME = 'vwc-note';
 
@@ -25,6 +27,6 @@ describe('note a11y', () => {
 		);
 		await waitNextTask();
 
-		await runAxeCore(actualElement);
+		await expect(actualElement).shadowDom.to.be.accessible();
 	});
 });