+++ components/select/src/vwc-select.ts
@@ -10,7 +10,7 @@ import {
 	PropertyValues,
 } from 'lit-element';
 import { Select as MWCSelect } from '@material/mwc-select';
-import { style as styleCoupling } from '@vonage/vvd-style-coupling/vvd-style-coupling.css.js';
+import { style as styleCoupling } from '@vonage/vvd-style-coupling/mdc-vvd-coupling.css';
 import { style as vwcSelectStyle } from './vwc-select.css';
 import { style as mwcSelectStyle } from '@material/mwc-select/mwc-select-css.js';
 import { associateWithForm } from '@vonage/vvd-foundation/form-association';
@@ -22,6 +22,8 @@ declare global {
 		'vwc-select': VWCSelect;
 	}
 }
+
+const DROPDOWN_ICON_CLASS = 'vvd-select-dropdown-icon';
 
 /* eslint-disable @typescript-eslint/ban-ts-comment */
 // @ts-ignore
@@ -55,7 +57,7 @@ export class VWCSelect extends MWCSelect {
 
 	async firstUpdated(): Promise<void> {
 		await super.firstUpdated();
-		this.replaceIcon();
+		this.replaceDropDownIcon();
 		associateWithForm(this, this.formElement);
 		handleAutofocus(this);
 	}
@@ -95,12 +97,11 @@ export class VWCSelect extends MWCSelect {
 		>`;
 	}
 
-	private replaceIcon(): void {
-		const ddIconClass = 'mdc-select__dropdown-icon';
-		const chevronIcon = document.createElement('vwc-icon');
-		chevronIcon.classList.add(ddIconClass);
-		chevronIcon.setAttribute('type', 'down');
-		this.shadowRoot?.querySelector(`.${ddIconClass}`)?.replaceWith(chevronIcon);
+	protected renderLeadingIcon(): TemplateResult | string {
+		if (!this.icon) {
+			return '';
+		}
+		return html`<vwc-icon class="vvd-select-icon" type="${this.icon}"></vwc-icon>`;
 	}
 
 	protected renderOutline(): TemplateResult | Record<string, unknown> {
@@ -111,5 +112,12 @@ export class VWCSelect extends MWCSelect {
 		return html`<vwc-notched-outline class="mdc-notched-outline vvd-notch">
 			${this.renderLabel()}
 		</vwc-notched-outline>`;
+	}
+
+	private replaceDropDownIcon(): void {
+		const chevronIcon = document.createElement('vwc-icon');
+		chevronIcon.classList.add(DROPDOWN_ICON_CLASS);
+		chevronIcon.setAttribute('type', 'down');
+		this.shadowRoot?.querySelector('.mdc-select__dropdown-icon')?.replaceWith(chevronIcon);
 	}
 }