+++ components/switch/stories/switch-basic.stories.js
@@ -4,7 +4,7 @@ import { spread } from '@open-wc/lit-helpers';
 import { argTypes } from './arg-types.js';
 
 export default {
-	title: 'Components/Atoms/Switch',
+	title: 'Components/Switch',
 	component: 'vwc-switch',
 	argTypes
 }